# Clientory App/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type EntitlementErrorResponseEntitlementCode =
  (typeof EntitlementErrorResponseEntitlementCode)[keyof typeof EntitlementErrorResponseEntitlementCode];

export const EntitlementErrorResponseEntitlementCode = {
  FREE_REPORT_ALREADY_USED: "FREE_REPORT_ALREADY_USED",
  SUBSCRIPTION_REQUIRED: "SUBSCRIPTION_REQUIRED",
  BUSINESS_PLAN_LIMIT_REACHED: "BUSINESS_PLAN_LIMIT_REACHED",
  SUBSCRIPTION_INACTIVE: "SUBSCRIPTION_INACTIVE",
  BILLING_PERIOD_ENDED: "BILLING_PERIOD_ENDED",
} as const;

export interface EntitlementErrorResponse {
  error: string;
  entitlementCode: EntitlementErrorResponseEntitlementCode;
  periodEnd?: string | null;
  scansUsed?: number | null;
  scansLimit?: number | null;
}

export interface RateLimitResponse {
  error: string;
  retryAfter?: number;
}

export interface SendOtpBody {
  email: string;
  turnstileToken?: string;
}

export interface SendOtpResponse {
  success: boolean;
  message: string;
}

export interface VerifyOtpBody {
  email: string;
  code: string;
}

export interface VerifyOtpResponse {
  success: boolean;
  emailToken: string;
  userId?: string | null;
}

export interface CreateScanBody {
  businessName: string;
  businessType: string;
  location: string;
  description?: string;
  website?: string;
  businessId: string;
}

export type ScanStatus = (typeof ScanStatus)[keyof typeof ScanStatus];

export const ScanStatus = {
  pending: "pending",
  generating_prompts: "generating_prompts",
  scanning: "scanning",
  completed: "completed",
  failed: "failed",
} as const;

export interface Scan {
  id: string;
  businessName: string;
  businessType: string;
  location: string;
  description?: string | null;
  website?: string | null;
  status: ScanStatus;
  score?: number | null;
  createdAt: string;
  businessId?: string | null;
  isFreeReport?: boolean;
}

export type CachedScanStatus =
  (typeof CachedScanStatus)[keyof typeof CachedScanStatus];

export const CachedScanStatus = {
  pending: "pending",
  generating_prompts: "generating_prompts",
  scanning: "scanning",
  completed: "completed",
  failed: "failed",
} as const;

export interface CachedScan {
  id: string;
  businessName: string;
  businessType: string;
  location: string;
  description?: string | null;
  website?: string | null;
  status: CachedScanStatus;
  score?: number | null;
  createdAt: string;
  cached: boolean;
}

export interface ScanPrompt {
  id: string;
  scanId: string;
  prompt: string;
  category: string;
  audience?: "individual" | "business" | null;
}

export type ScanResultProvider =
  (typeof ScanResultProvider)[keyof typeof ScanResultProvider];

export const ScanResultProvider = {
  openai: "openai",
  anthropic: "anthropic",
  gemini: "gemini",
} as const;

export interface ScanResult {
  id: string;
  scanPromptId: string;
  provider: ScanResultProvider;
  response: string;
  mentioned: boolean;
  createdAt: string;
}

export type ScanDetailPromptsItem = {
  prompt: ScanPrompt;
  results: ScanResult[];
};

export interface ScanDetail {
  scan: Scan;
  prompts: ScanDetailPromptsItem[];
  recommendations: string[];
}

export interface FirmLocation {
  city: string;
  state: string;
  neighborhood?: string | null;
  isHQ: boolean;
}

export interface FirmPartner {
  name: string;
  title: string;
  trackIndependently: boolean;
}

export interface FirmCompetitor {
  name: string;
  location?: string | null;
}

export type CreateBusinessBodyRankingsItem = {
  source: string;
  category: string;
  year: number;
};

export interface CreateBusinessBody {
  name?: string;
  website?: string | null;
  businessType?: string;
  location?: string;
  description?: string | null;
  legalName?: string | null;
  brandName?: string | null;
  firmType?: string | null;
  yearFounded?: number | null;
  partners?: FirmPartner[] | null;
  locations?: FirmLocation[] | null;
  geographicScope?: string | null;
  primaryServices?: string[] | null;
  deliverables?: string[] | null;
  specialties?: string[] | null;
  industriesServed?: string[] | null;
  clientStages?: string[] | null;
  decisionMakers?: string[] | null;
  directCompetitors?: FirmCompetitor[] | null;
  rankings?: CreateBusinessBodyRankingsItem[] | null;
  topGSCQueries?: string[] | null;
  clientType?: string | null;
  individualServices?: string[] | null;
  individualDeliverables?: string[] | null;
  individualSpecialties?: string[] | null;
  businessServices?: string[] | null;
  businessDeliverables?: string[] | null;
  businessSpecialties?: string[] | null;
}

export type BusinessSubscriptionInfoStatus =
  (typeof BusinessSubscriptionInfoStatus)[keyof typeof BusinessSubscriptionInfoStatus];

export const BusinessSubscriptionInfoStatus = {
  active: "active",
  past_due: "past_due",
  canceled: "canceled",
  incomplete: "incomplete",
} as const;

export interface BusinessSubscriptionInfo {
  status: BusinessSubscriptionInfoStatus;
  currentPeriodEnd: string;
  cancelAtPeriodEnd: boolean;
}

export interface BusinessUsageInfo {
  scansUsed: number;
  scansLimit: number;
  periodEnd: string;
}

export interface Business {
  id: string;
  name: string;
  website?: string | null;
  businessType: string;
  location: string;
  description?: string | null;
  legalName?: string | null;
  brandName?: string | null;
  firmType?: string | null;
  yearFounded?: number | null;
  locations?: unknown | null;
  primaryServices?: unknown | null;
  deliverables?: unknown | null;
  specialties?: unknown | null;
  industriesServed?: unknown | null;
  clientStages?: unknown | null;
  decisionMakers?: unknown | null;
  directCompetitors?: unknown | null;
  clientType?: string | null;
  individualServices?: unknown | null;
  individualDeliverables?: unknown | null;
  individualSpecialties?: unknown | null;
  businessServices?: unknown | null;
  businessDeliverables?: unknown | null;
  businessSpecialties?: unknown | null;
  profileUpdatedAt?: string | null;
  hasPromptSet: boolean;
  createdAt: string;
  subscription?: BusinessSubscriptionInfo | null;
  usage?: BusinessUsageInfo | null;
}

export interface BusinessListResponse {
  businesses: Business[];
}

export interface SubscribeBody {
  businessId: string;
}

export type BillingRedirectResponseType =
  (typeof BillingRedirectResponseType)[keyof typeof BillingRedirectResponseType];

export const BillingRedirectResponseType = {
  checkout: "checkout",
  portal: "portal",
} as const;

export interface BillingRedirectResponse {
  url: string;
  type: BillingRedirectResponseType;
}

export interface BillingPortalResponse {
  url: string;
}

export interface BillingStatus {
  businessId: string;
  isFree: boolean;
  isPaid: boolean;
  subscription?: BusinessSubscriptionInfo | null;
  usage?: BusinessUsageInfo | null;
}

export type PromptItemSourceVariables = { [key: string]: unknown };

export type PromptItemAudience =
  | (typeof PromptItemAudience)[keyof typeof PromptItemAudience]
  | null;

export const PromptItemAudience = {
  individual: "individual",
  business: "business",
} as const;

export interface PromptItem {
  id: string;
  category: string;
  text: string;
  sourceVariables: PromptItemSourceVariables;
  audience?: PromptItemAudience;
}

export interface PromptSet {
  id: string;
  businessId: string;
  generatedAt: string;
  prompts: PromptItem[];
  isLocked: boolean;
  substitutionNotes?: string | null;
  createdAt: string;
}

export interface SuggestCompetitorsBody {
  firmType: string;
  city: string;
  state: string;
  primaryServices?: string[];
}

export type SuggestCompetitorsResponseSuggestionsItem = {
  name: string;
  location?: string | null;
};

export interface SuggestCompetitorsResponse {
  suggestions: SuggestCompetitorsResponseSuggestionsItem[];
}