# Clientory App/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP verification code
      description: Sends a 6-digit verification code to the provided email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpBody"
      responses:
        "200":
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendOtpResponse"
        "400":
          description: Invalid input or CAPTCHA failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpBody"
      responses:
        "200":
          description: OTP verified, email token returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyOtpResponse"
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /billing/subscribe:
    post:
      operationId: subscribeBusiness
      tags: [billing]
      summary: Create a Stripe Checkout or Portal session for a business
      parameters:
        - name: x-email-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeBody"
      responses:
        "200":
          description: Checkout or portal URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingRedirectResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /billing/portal:
    post:
      operationId: createBillingPortal
      tags: [billing]
      summary: Create a Stripe Customer Portal session
      parameters:
        - name: x-email-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Portal URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingPortalResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /billing/status/{businessId}:
    get:
      operationId: getBillingStatus
      tags: [billing]
      summary: Get billing status for a business
      parameters:
        - name: x-email-token
          in: header
          required: true
          schema:
            type: string
        - name: businessId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Billing status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingStatus"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Business not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RateLimitResponse:
      type: object
      properties:
        error:
          type: string
        retryAfter:
          type: number
      required:
        - error
    SendOtpBody:
      type: object
      properties:
        email:
          type: string
          format: email
        turnstileToken:
          type: string
      required:
        - email
    SendOtpResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    VerifyOtpBody:
      type: object
      properties:
        email:
          type: string
          format: email
        code:
          type: string
      required:
        - email
        - code
    VerifyOtpResponse:
      type: object
      properties:
        success:
          type: boolean
        emailToken:
          type: string
        userId:
          type: string
          nullable: true
      required:
        - success
        - emailToken
...
    BusinessSubscriptionInfo:
      type: object
      properties:
        status:
          type: string
          enum: [active, past_due, canceled, incomplete]
        currentPeriodEnd:
          type: string
          format: date-time
        cancelAtPeriodEnd:
          type: boolean
      required:
        - status
        - currentPeriodEnd
        - cancelAtPeriodEnd
    BusinessUsageInfo:
      type: object
      properties:
        scansUsed:
          type: number
        scansLimit:
          type: number
        periodEnd:
          type: string
          format: date-time
      required:
        - scansUsed
        - scansLimit
        - periodEnd
...
    SubscribeBody:
      type: object
      properties:
        businessId:
          type: string
      required:
        - businessId
    BillingRedirectResponse:
      type: object
      properties:
        url:
          type: string
        type:
          type: string
          enum: [checkout, portal]
      required:
        - url
        - type
    BillingPortalResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    BillingStatus:
      type: object
      properties:
        businessId:
          type: string
        isFree:
          type: boolean
        isPaid:
          type: boolean
        subscription:
          $ref: "#/components/schemas/BusinessSubscriptionInfo"
          nullable: true
        usage:
          $ref: "#/components/schemas/BusinessUsageInfo"
          nullable: true
      required:
        - businessId
        - isFree
        - isPaid
...