# Clientory App/lib/api-zod/src/generated/types/businessUsageInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessUsageInfo {
  scansUsed: number;
  scansLimit: number;
  periodEnd: Date;
}